Add typings for change_trust.
